--- shared/src/commonMain/kotlin/me/user/shared/Greeting.kt
@@ -1,8 +1,0 @@
-package me.user.shared
-
-
-class Greeting {
-    fun greeting(): String {
-        return "Hello, ${Platform().platform}!"
-    }
-}
