--- shared/src/commonMain/kotlin/me/user/shared/Platform.kt
@@ -1,5 +1,0 @@
-package me.user.shared
-
-expect class Platform() {
-    val platform: String
-}
