--- shared/src/iosMain/kotlin/me/user/shared/Platform.kt
@@ -1,8 +1,0 @@
-package me.user.shared
-
-
-import platform.UIKit.UIDevice
-
-actual class Platform actual constructor() {
-    actual val platform: String = UIDevice.currentDevice.systemName() + " " + UIDevice.currentDevice.systemVersion
-}
