--- shared/src/iosMain/kotlin/me/user/shared/SwiftHelper.kt
@@ -1,5 +1,0 @@
-package me.user.shared
-
-class SwiftHelper {
-    fun getViewController() = getSkikoViewContoller()
-}
