+++ shared/src/iosMain/kotlin/org/jetbrains/skiko/sample/SwiftHelper.kt
@@ -1,0 +1,7 @@
+package org.jetbrains.skiko.sample
+
+import org.jetbrains.skiko.sample.getSkikoViewContoller
+
+class SwiftHelper {
+    fun getViewController() = getSkikoViewContoller()
+}
