--- shared/src/iosTest/kotlin/me/user/shared/iosTest.kt
@@ -1,12 +1,0 @@
-package me.user.shared
-
-import kotlin.test.Test
-import kotlin.test.assertTrue
-
-class IosGreetingTest {
-
-    @Test
-    fun testExample() {
-        assertTrue(Greeting().greeting().contains("iOS"), "Check iOS is mentioned")
-    }
-}
